pyproject.toml
[tool.poetry]
name = "comtec_test"
version = "0.1.0"
description = ""
authors = ["vildan-valeev <vildan.maydanovich.valeev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
environs = "^9.5.0"
django-split-settings = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
